requirements.txt
pandas==2.2.2
matplotlib==3.10.0
python-dotenv==1.0.1
openpyxl==3.1.2